Refactor: make api call to rickandmorty api endpoint from within the <CharacterList /> component instead of <App/>
