+++ src/components/CharacterList.js
@@ -1,12 +1,22 @@
 import React, { useEffect, useState } from "react";
+import axios from "axios";
+
+
+const initialCharacterList = []
+const url = "https://rickandmortyapi.com/api/character/";
 
 export default function CharacterList() {
-  // TODO: Add useState to track data from useEffect
+  const [charactersList, setCharactersList] = useState(initialCharacterList)
 
   useEffect(() => {
-    // TODO: Add API Request here - must run in `useEffect`
-    //  Important: verify the 2nd `useEffect` parameter: the dependancies array!
-  }, []);
+    axios.get(url)
+    .then(res => {
+      setCharactersList(res.data.results);
+    })
+    .catch(err => {
+      console.log(err.message);
+    })
+  }, [])
 
   return (
     <section className="character-list">
